+++ monolith-api/src/main/scala/gg/warcraft/monolith/api/player/statistic/Statistics.scala
@@ -1,7 +1,7 @@
 package gg.warcraft.monolith.api.player.statistic
 
 case class Statistics private (statistics: Map[String, Statistic]) {
-  def get(statistic: String): Int =
+  def get(statistic: String): Long =
     statistics
       .get(statistic)
       .map { _.value }
